feat-refactoring all providers into a single script
